Add Feedback combobox questions #670
Started a quick mockup of how this could be done. This adds the feature
in the model and form and exposes the results in
http://..../results/chartdata/

Pretty much everything sort of works, but has a wonky admin interface as
we cannot nest InlineModelAdmin rows.

The new methodnames in models.py should be cleaned up before a possible
merge. And some way of showing the results is needed.