Change how anonymus voter hashing works
The anonymous voter hash that is stored in the database is now hashed with the username, while the cookie hash remains the same.
Everytime the anonymous voter object is looked up in the database it will have to be hashed with the username.
This is to prevent cookie stealing, which could make it possible for a user to vote on behalf of other users.