We can't have nice things and bs3
The `$(this).tab('show');` in `assets/frontpage/initFrontpage.js` was
broken, since `.tab` was not defined. `.tab` is a bootstrap3-extension
to jQuery https://getbootstrap.com/docs/3.4/javascript/#tabs. To make
that extension method available you need to rely on the side-effects
of `import 'bootstrap'`, which only applies it to the _global_ $/jQuery.

When you `import $ from 'jquery'`, that creates a local and specifically
non-global `$`. Generally speaking you would then immediately set your
`$` to be the global `$` by using `globalThis.jQuery = globalThis.$ =
$`. We import and thus define the global `$` from a `<script>`-element
in the base HTML template, and thus this causes a "conflict" between the
esm/esbuild `$` and the already imported one.

In particular `import 'bootstrap'` only applies the extensions to the
global one. Ideally we would just get rid of this use of global
variables (and jQuery as a whole). For now just give up on attempting to
use jQuery from npm, and go back to jQuery 2.X from the Django template.

The strange thing here is probably rather that there were imports in the
webpack setup, since it was actually threated as a global variable, thus
the imports previously did "nothing".