Remove bootstrap-datetimepicker
Similar to 6735dda, it also used jQuery-extension-methods, but with
the datetimepicker I could not determine which jQuery it was modifying.
Thanks to newer input-types, we can rely on browser-support for these inputs,
is a much better experience, in particular for mobile users.

This also removes a `HTML5RequiredMixin`, which AFAIK has been redundant
since Django 1.10

It is very likely there are leftovers that have not been removed