Fix double-counting weight, fewer unneeded marks
`sanction_mark` was double counting the mark-weight since it added for
the specific mark _and_ used user_weight after the mark user was added
to those affected.

Skip save-ing marks being sent to user_sanctions, user_sanctions saves
them themselves.

MarkUser _still_ counts both user_weight and mark weight since it
assumes the user has not yet been added to the list of mark's users.
It is intended to trigger from changes in the admin-panel, so that is
_fine_.

Move mark-creation to only when necessary